Moved around some final pieces, module-ized things
