+++ config/initializers/setup_services.rb
@@ -1,0 +1,1 @@
+Github::Auth.setup
